Refactored to build against JDK 6
In order to build the plugin for AppCode it's necessary to build against JDK 6. Therefore some parts of the plugin had to be changed. Mainly the previous file watcher implementation had to be changed to a VirtualFileAdapter. Unfortunately the current FileListener refreshes only when the IDE comes into foreground again. This has to be fixed.